+++ openslides/old_urls.py
@@ -3,28 +3,22 @@ from django.conf.urls import include, url
 from django.contrib.staticfiles.urls import urlpatterns
 from django.views.generic import RedirectView
 
+from openslides.core import views as core_views
 from openslides.mediafiles.views import protected_serve
+from openslides.utils.plugins import get_all_plugin_urlpatterns
 from openslides.utils.rest_api import router
 
-from .core import views as core_views
 
-
-# Urls for /static/ are already in urlpatterns
+urlpatterns += get_all_plugin_urlpatterns()
 
 urlpatterns += [
-    # URLs for /media/
     url(r'^%s(?P<path>.*)$' % settings.MEDIA_URL.lstrip('/'), protected_serve, {'document_root': settings.MEDIA_ROOT}),
-
-    # When a url without a leading slash is requested, redirect to the url with
-    # the slash. This line has to be after static and media files.
     url(r'^(?P<url>.*[^/])$', RedirectView.as_view(url='/%(url)s/', permanent=True)),
-
-    # URLs for the rest system
     url(r'^rest/', include(router.urls)),
-
-    # Other urls defined by modules and plugins
-    url(r'^apps/', include('openslides.urls_apps')),
-
+    url(r'^agenda/', include('openslides.agenda.urls')),
+    url(r'^motions/', include('openslides.motions.urls')),
+    url(r'^users/', include('openslides.users.urls')),
+    url(r'^core/', include('openslides.core.urls')),
     # The old angular webclient
     # TODO: Change me or at least my comment
     url(r'^webclient/(?P<realm>site|projector)/$',